requirements.txt
oci[core]
paramiko
PyYAML
requests
qrcode